+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Adds a duplicate key checker to the YAML `Loader` that raises an error when a duplicate key is
   found, and points to the file and line number that caused the error. The YAML `Loader` modification
   maintains compliance with the existing JSON validation protocols.
+- Added a refactored iron electrowinning model with performance and cost models based on recent literature from Humbert and Stinn
 
 ## 0.6 [February 10, 2026]